+++ app/controllers/api/v1/reviews_controller.rb
@@ -1,15 +1,15 @@
 class Api::V1::ReviewsController < ApplicationController
-  protect_from_forgery unless: -> { request.format.json? }
 
   def create
     restaurant = Restaurant.find(params[:restaurant_id])
     review = Review.new(review_params)
     review.restaurant = restaurant
+    review.user = current_user
 
     if review.save
       render json: { review: review }
     else
-      render json: { error: review.errors.full_messages }
+      render json: { error: review.errors.full_messages }, status: 400
     end
   end
